Rename insertion_ordered_map module to map, because rust-analyzer keeps thinking it doesn't exist.
Also remove some unnecessary mutables. (Warnings were overlooked because of rust-analyzer wigging out.)
