+++ crates/common/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod error;
-pub mod insertion_ordered_map;
+pub mod map;
 pub mod name;
 pub mod numbers;
 pub mod traits;
